refactor(spec): separate EAGLE state from tree utilities
outline:
- PR role — cleanup: Give EAGLE state contracts and tree helpers distinct module boundaries.
- Sequence — Builds on the reduced EAGLE runtime surface and enables focused Qwen3-8B optimization work.

modify:
- Before -> after: Move EagleDraftInput, EagleVerifyInput, pytree metadata helpers, batch transitions, and verification dispatch from eagle_util.py into eagle_info.py.
- Redirect runtime, scheduler, attention, test, and documentation references while leaving tree construction and shared KV mapping helpers in eagle_util.py.
- Scope boundary: Preserve the state-class AST and all EAGLE tree, sampling, overlap, and DP behavior.

reason:
- Problem: A nearly thousand-line utility module mixed persistent state contracts with tree algorithms, obscuring ownership and review boundaries.
- Design: Mirror the existing speculative info-module organization and keep the move behavior-preserving with focused pytree and DP regression coverage.
