+++ python/sgl_jax/srt/speculative/eagle_info.py
@@ -1,0 +1,725 @@
+"""EAGLE speculative-decoding state objects and their batch transitions."""
+
+from __future__ import annotations
+
+import copy
+from dataclasses import dataclass
+from typing import TYPE_CHECKING, Any, ClassVar
+
+import jax
+import jax.numpy as jnp
+import numpy as np
+from flax import nnx
+from jax.sharding import Mesh, NamedSharding
+from jax.sharding import PartitionSpec as P
+from jax.tree_util import register_pytree_node_class
+
+from sgl_jax.srt.kernels.speculative.kernel import (
+    top_k_renorm_prob,
+    top_p_renorm_prob,
+    tree_speculative_sampling_target_only,
+)
+from sgl_jax.srt.kernels.speculative.verify_tree_greedy_kernel import verify_tree_greedy
+from sgl_jax.srt.layers.logits_processor import LogitsProcessorOutput
+from sgl_jax.srt.managers.schedule_batch import (
+    ModelWorkerBatch,
+    ScheduleBatch,
+    get_last_loc,
+    global_server_args_dict,
+)
+from sgl_jax.srt.mem_cache.common import (
+    alloc_paged_token_slots_extend,
+    alloc_token_slots,
+)
+from sgl_jax.srt.model_executor.forward_batch_info import CaptureHiddenMode, ForwardMode
+from sgl_jax.srt.speculative.eagle_util import assign_req_to_token_pool
+from sgl_jax.srt.speculative.overlap_utils import use_legacy_eagle3_non_overlap
+from sgl_jax.srt.speculative.spec_utils import (
+    SIMULATED_ACCEPTANCE_CONFIG,
+    apply_simulated_acceptance,
+)
+
+if TYPE_CHECKING:
+    from sgl_jax.srt.managers.scheduler import GenerationBatchResult
+
+
+def _is_jax_leaf(value: Any) -> bool:
+    """Detect sentinel nodes generated by jax.tree_util when shaping pytrees."""
+    cls = value.__class__
+    return cls.__name__ == "Leaf" and cls.__module__.startswith("jax.")
+
+
+def _as_int32_array(value: Any, *, fallback: int = -1) -> Any:
+    """Convert scalar-like metadata into int32 arrays without forcing device work."""
+    if value is None:
+        return None
+    if isinstance(value, jax.Array):
+        return value
+    if isinstance(value, np.ndarray):
+        return np.asarray(value, dtype=np.int32)
+    if isinstance(value, (int, np.integer)):
+        return np.asarray(int(value), dtype=np.int32)
+    if isinstance(value, (list, tuple)):
+        return np.asarray(value, dtype=np.int32)
+    if _is_jax_leaf(value):
+        return np.asarray(fallback, dtype=np.int32)
+    try:
+        return np.asarray(value, dtype=np.int32)
+    except (TypeError, ValueError) as exc:
+        raise TypeError(
+            f"Unable to convert value of type {type(value)} into int32 metadata array."
+        ) from exc
+
+
+@register_pytree_node_class
+@dataclass
+class EagleDraftInput:
+    """Next-round draft state — the only persistent cross-round spec state.
+
+    MUST NOT hold worker/runner/pool/future handles.
+    Under DP (Route 1), per-request fields use DP-padded order.
+    """
+
+    ALLOC_LEN_PER_DECODE: ClassVar[int] = None
+
+    # --- Cross-round draft state (device arrays, consumed by next draft) ---
+    #: device ``(b, topk)`` — top-k probs from previous draft/draft_extend.
+    topk_p: jax.Array | None = None
+    #: device ``(b, topk)`` — top-k token ids.
+    topk_index: jax.Array | None = None
+    #: device ``(b, hidden_size)`` — minimal hidden state for next draft step.
+    #: Multi-layer MTP keeps per-step hidden locally inside one
+    #: ``MultiLayerDraftWorker.draft()``; only this cross-round slice persists.
+    hidden_states: jax.Array | None = None
+    #: static metadata (pytree aux); changing it triggers a new compile shape.
+    capture_hidden_mode: CaptureHiddenMode = CaptureHiddenMode.FULL
+
+    # --- Draft-extend inputs (device unless ``_cpu`` suffixed) ---
+    #: device ``(b,)`` — verified token starting the next draft.
+    verified_id: jax.Array | None = None
+    #: device ``(b,)`` — accepted length used to select hidden in draft-extend.
+    accept_length: jax.Array | None = None
+    #: host ``(b,)`` int32 mirror of ``accept_length`` for scheduler bookkeeping.
+    accept_length_cpu: np.ndarray | None = None
+
+    # --- Attention-backend metadata (host, participates in metadata build) ---
+    kv_indices: np.ndarray | None = None
+
+    # --- Padding shape (static; participates in JIT cache key) ---
+    num_tokens_per_batch: int = -1
+    num_tokens_for_logprob_per_batch: int = -1
+
+    # --- KV lifetime (host, scheduler-visible) ---
+    #: host ``(b,)`` — KV length already allocated in ``req_to_token_pool`` for
+    #: next-round pre-allocation and over-allocated slot release. Distinct from
+    #: ``accept_length`` (logical) and ``new_seq_lens`` (scheduler-visible).
+    allocate_lens: np.ndarray | None = None
+    #: host ``(b,)`` — scheduler-visible logical length after verify. May be
+    #: derived from ``old_seq_lens + accept_length`` if not stored.
+    new_seq_lens: np.ndarray | None = None
+    #: host ``(b,)`` — req_pool_indices used to gather relay-buffer state.
+    future_indices: np.ndarray | None = None
+
+    def new_tokens_required_next_decode(self, requests, page_size: int) -> int:
+        target_alloc = self.ALLOC_LEN_PER_DECODE * 2
+        total = 0
+        for req in requests:
+            cur = req.kv_allocated_len
+            nxt = max(cur, req.kv_committed_len + target_alloc)
+            total += ((nxt + page_size - 1) // page_size) * page_size - (
+                (cur + page_size - 1) // page_size
+            ) * page_size
+        return total
+
+    def tree_flatten(self):
+        accept_length_cpu_arr = (
+            np.empty((0,), dtype=np.int32)
+            if self.accept_length_cpu is None
+            else _as_int32_array(self.accept_length_cpu, fallback=0)
+        )
+
+        num_tokens_per_batch_arr = _as_int32_array(self.num_tokens_per_batch)
+        num_tokens_for_logprob_arr = _as_int32_array(self.num_tokens_for_logprob_per_batch)
+
+        children = (
+            self.topk_p,
+            self.topk_index,
+            self.hidden_states,
+            self.verified_id,
+            self.accept_length,
+            self.kv_indices,
+            self.future_indices,
+            accept_length_cpu_arr,
+            num_tokens_per_batch_arr,
+            num_tokens_for_logprob_arr,
+        )
+
+        aux_data = {
+            "capture_hidden_mode": self.capture_hidden_mode,
+        }
+        return (children, aux_data)
+
+    @classmethod
+    def tree_unflatten(cls, aux_data, children):
+        obj = cls.__new__(cls)
+        obj.capture_hidden_mode = aux_data["capture_hidden_mode"]
+        obj.topk_p = children[0]
+        obj.topk_index = children[1]
+        obj.hidden_states = children[2]
+        obj.verified_id = children[3]
+        obj.accept_length = children[4]
+        obj.kv_indices = children[5]
+        obj.future_indices = children[6]
+
+        obj.accept_length_cpu = children[7]
+        obj.num_tokens_per_batch = children[8]
+        obj.num_tokens_for_logprob_per_batch = children[9]
+
+        return obj
+
+    def prepare_for_extend_after_target_prefill(self, model_worker_batch: ModelWorkerBatch):
+        # Prefill only generate 1 token.
+        assert self.verified_id.shape[0] == model_worker_batch.real_bs, (
+            f"{self.verified_id.shape=} {model_worker_batch.real_bs=}"
+        )
+
+        # Walk the DP-padded layout in (rank, slot) order so token offsets line
+        # up with mwb.input_ids' token-major DP layout. Iterating real_bs and
+        # indexing extend_seq_lens by `i` skipped real reqs whose padded slot
+        # index > i (e.g. dp>1 prefill with only rank>0 active).
+        dp_size = model_worker_batch.dp_size
+        # per_dp_bs_size already covers both dp=1 (= total_bs) and dp>1
+        # (= total_bs / dp). Hardcoding 1 here drops dp=1 multi-req.
+        per_dp_bs = model_worker_batch.per_dp_bs_size
+        total_tok = len(model_worker_batch.input_ids)
+        per_dp_tok = total_tok // dp_size
+        extend_seq_lens = model_worker_batch.extend_seq_lens
+        flat_idx = 0  # index into self.verified_id (cross-rank flat)
+        for dp_rank in range(dp_size):
+            pt = dp_rank * per_dp_tok
+            for slot_in_rank in range(per_dp_bs):
+                slot = dp_rank * per_dp_bs + slot_in_rank
+                extend_len = int(extend_seq_lens[slot])
+                if extend_len == 0:
+                    continue
+                input_ids = model_worker_batch.input_ids[pt : pt + extend_len]
+                model_worker_batch.input_ids[pt : pt + extend_len] = np.concatenate(
+                    (input_ids[1:], self.verified_id[flat_idx].reshape(1))
+                )
+                pt += extend_len
+                flat_idx += 1
+        assert flat_idx == model_worker_batch.real_bs, (
+            f"prepare_for_extend_after_target_prefill mismatch: walked {flat_idx} "
+            f"real reqs but real_bs={model_worker_batch.real_bs}"
+        )
+
+    def prepare_for_extend_after_verify(
+        self,
+        model_worker_batch: ModelWorkerBatch,
+        draft_model_runner: Any,
+        batch_output: GenerationBatchResult,
+        speculative_num_draft_tokens: int,
+    ):
+        legacy_non_overlap = (
+            model_worker_batch.spec_algorithm is not None
+            and model_worker_batch.spec_algorithm.is_eagle3()
+            and getattr(batch_output.next_draft_input, "future_indices", None) is None
+        )
+        model_worker_batch.spec_info_padded = self
+        sel = model_worker_batch.logits_indices_selector
+        model_worker_batch.seq_lens[sel] = (
+            model_worker_batch.seq_lens[sel] + speculative_num_draft_tokens - 1
+        )
+        bs = batch_output.accept_lens.shape[0]
+        step_plus_1 = model_worker_batch.input_ids.shape[0] // bs
+        positions = getattr(batch_output.next_draft_input, "positions", None)
+        if positions is not None and not legacy_non_overlap:
+            model_worker_batch.positions = positions
+        model_worker_batch.extend_seq_lens = np.zeros((bs,), dtype=np.int32)
+        model_worker_batch.extend_seq_lens[sel] = step_plus_1
+        # Per-rank-local cumsum: _select_hidden_states is a shard_map rank-local
+        # gather, so indices must be offsets into each rank's own hidden shard.
+        dp = model_worker_batch.dp_size
+        per_dp = model_worker_batch.per_dp_bs_size if dp > 1 else bs
+        model_worker_batch.logits_indices = (
+            model_worker_batch.extend_seq_lens.reshape(dp, per_dp)
+            .cumsum(axis=1, dtype=np.int32)
+            .ravel()
+            - 1
+        )
+        model_worker_batch.capture_hidden_mode = CaptureHiddenMode.FULL
+        model_worker_batch.spec_info_padded.capture_hidden_mode = CaptureHiddenMode.FULL
+        model_worker_batch.forward_mode = ForwardMode.DRAFT_EXTEND
+        model_worker_batch.spec_info_padded.hidden_states = (
+            batch_output.next_draft_input.hidden_states
+        )
+        if legacy_non_overlap or model_worker_batch.spec_info_padded.accept_length is None:
+            model_worker_batch.spec_info_padded.accept_length = batch_output.accept_lens
+        model_worker_batch.input_ids = batch_output.next_draft_input.verified_id
+        forward_metadata = draft_model_runner.attn_backend.get_eagle_forward_metadata(
+            model_worker_batch
+        )
+
+        draft_model_runner.attn_backend.forward_metadata = forward_metadata
+        from sgl_jax.srt.layers.logits_processor import LogitsMetadata
+        from sgl_jax.srt.utils.jax_utils import device_array
+
+        if legacy_non_overlap or model_worker_batch.return_logprob:
+            logits_metadata = LogitsMetadata.from_model_worker_batch(
+                model_worker_batch, draft_model_runner.mesh
+            )
+        else:
+            sharding = NamedSharding(draft_model_runner.mesh, P("data"))
+
+            def _to_device(value):
+                if value is None:
+                    return None
+                if isinstance(value, jax.Array):
+                    current_sharding = value.sharding
+                    if current_sharding == sharding:
+                        return value
+                    return jax.device_put(value, sharding)
+                return device_array(value, sharding=sharding)
+
+            extend_seq_lens_for_logits = getattr(
+                model_worker_batch.spec_info_padded,
+                "extend_seq_lens_for_draft_extend",
+                None,
+            )
+            if extend_seq_lens_for_logits is None:
+                extend_seq_lens_for_logits = model_worker_batch.extend_seq_lens
+            logits_indices_for_logits = getattr(
+                model_worker_batch.spec_info_padded,
+                "logits_indices_for_draft_extend",
+                None,
+            )
+            if logits_indices_for_logits is None:
+                logits_indices_for_logits = model_worker_batch.logits_indices
+
+            logits_metadata = LogitsMetadata(
+                forward_mode=model_worker_batch.forward_mode,
+                capture_hidden_mode=model_worker_batch.capture_hidden_mode,
+                extend_return_logprob=False,
+                extend_return_top_logprob=False,
+                extend_token_ids_logprob=False,
+                extend_seq_lens=_to_device(extend_seq_lens_for_logits),
+                logits_indices=_to_device(logits_indices_for_logits),
+                accept_lens=_to_device(model_worker_batch.spec_info_padded.accept_length),
+                extend_seq_lens_cpu=None,
+                extend_logprob_start_lens_cpu=None,
+                extend_logprob_pruned_lens_cpu=None,
+                top_logprobs_nums=model_worker_batch.top_logprobs_nums,
+                token_ids_logprobs=model_worker_batch.token_ids_logprobs,
+                extend_input_logprob_token_ids_device=_to_device(
+                    model_worker_batch.extend_input_logprob_token_ids
+                ),
+            )
+        return model_worker_batch, logits_metadata
+
+    def prepare_for_decode(self, schedule_batch: ScheduleBatch):
+        # spec_info / allocate_lens are global (flat across DP ranks). Allocate
+        # per-rank (allocator + swa_mapping are per-dp_rank at dp>1); concat
+        # back to global-flat. (#1053 P1-5b — was dp_rank=0 for all, so rank>0
+        # reqs allocated from rank 0's pool / updated swa_mapping[0].)
+        bs = schedule_batch.batch_size()
+        assert self.allocate_lens.shape[0] == bs, (
+            f" {self.allocate_lens.shape[0]=} but batch_size is {bs} "
+        )
+        page_size = schedule_batch.token_to_kv_pool_allocator.page_size
+        new_alloc_chunks = []
+        flat_off = 0
+        legacy_non_overlap = use_legacy_eagle3_non_overlap(
+            schedule_batch.enable_overlap, schedule_batch.spec_algorithm
+        )
+        for dp_rank, info in enumerate(schedule_batch.reqs_info):
+            if info.seq_lens is None or len(info.seq_lens) == 0:
+                continue
+            bs_r = len(info.seq_lens)
+            if legacy_non_overlap:
+                seq_r = np.asarray(info.seq_lens)
+                new_r = seq_r + self.ALLOC_LEN_PER_DECODE - 1
+                old_r = self.allocate_lens[flat_off : flat_off + bs_r]
+            else:
+                old_r = np.asarray([req.kv_allocated_len for req in info.reqs], dtype=np.int32)
+                committed_r = np.asarray(
+                    [req.kv_committed_len for req in info.reqs], dtype=np.int32
+                )
+                new_r = np.maximum(old_r, committed_r + 2 * self.ALLOC_LEN_PER_DECODE)
+            ext_r = int((new_r - old_r).sum())
+            if page_size == 1:
+                ocl_r = alloc_token_slots(schedule_batch.tree_cache, ext_r, dp_rank=dp_rank)
+            else:
+                last_loc_r = get_last_loc(
+                    schedule_batch.req_to_token_pool.req_to_token,
+                    info.req_pool_indices,
+                    old_r,
+                )
+                ocl_r = alloc_paged_token_slots_extend(
+                    schedule_batch.tree_cache,
+                    old_r,
+                    new_r,
+                    last_loc_r,
+                    ext_r,
+                    dp_rank=dp_rank,
+                )
+            assign_req_to_token_pool(
+                info.req_pool_indices, schedule_batch.req_to_token_pool, old_r, new_r, ocl_r
+            )
+            new_alloc_chunks.append(new_r)
+            # Per-rank store (matches nospec extend); _get_spec_decode_mwb_dp
+            # DP-segments these so each rank's P("data") shard = its own slots.
+            info.out_cache_loc = np.asarray(ocl_r, dtype=np.int32)
+            for req, allocated_len in zip(info.reqs, new_r):
+                req.decode_batch_idx += 1
+                req.kv_allocated_len = int(allocated_len)
+                if not legacy_non_overlap:
+                    req.kv_committed_len += 1
+            flat_off += bs_r
+            info.seq_lens_sum = np.sum(info.seq_lens).item()
+
+        self.allocate_lens = np.concatenate(new_alloc_chunks)
+
+    def prepare_for_draft_decode(self, model_worker_batch: ModelWorkerBatch, topk: int):
+        self.capture_hidden_mode = CaptureHiddenMode.LAST
+        self.num_tokens_per_batch = topk
+        self.num_tokens_for_logprob_per_batch = topk
+        model_worker_batch.return_hidden_states = False
+        model_worker_batch.seq_lens_sum = np.sum(model_worker_batch.seq_lens)
+
+    def _ensure_host(self):
+        """Move device arrays to host (numpy) to avoid variable-shape device ops.
+
+        Spec decode postprocessing (filter_batch, merge_batch, split/concat)
+        operates on per-request arrays whose size varies with real_bs. Keeping
+        these as jax.Array triggers implicit JIT compilation for each new shape.
+        Converting to numpy first and re-uploading at bucket-padded size in
+        _scatter_spec_info_to_dp_slots eliminates those persistent cache misses.
+        """
+        device_fields = ("topk_p", "topk_index", "hidden_states", "verified_id", "accept_length")
+        to_copy = []
+        for f in device_fields:
+            v = getattr(self, f, None)
+            if v is not None and hasattr(v, "copy_to_host_async"):
+                jax.copy_to_host_async(v)
+                to_copy.append(f)
+        for f in to_copy:
+            setattr(self, f, np.asarray(getattr(self, f)))
+
+    def filter_batch(self, new_indices: np.ndarray, has_been_filtered: bool = True):
+        new_indices = np.asarray(new_indices)
+        if self.future_indices is not None:
+            src = np.asarray(self.future_indices)
+            idx = (
+                slice(0, len(new_indices))
+                if has_been_filtered and len(new_indices) == len(src)
+                else new_indices
+            )
+            self.future_indices = src[idx]
+            if self.allocate_lens is not None:
+                self.allocate_lens = np.asarray(self.allocate_lens)[idx]
+            if self.new_seq_lens is not None:
+                self.new_seq_lens = np.asarray(self.new_seq_lens)[idx]
+            if self.accept_length_cpu is not None:
+                self.accept_length_cpu = np.asarray(self.accept_length_cpu)[idx]
+            return
+
+        self._ensure_host()
+        if has_been_filtered and len(new_indices) == len(self.topk_p):
+            self.topk_p = self.topk_p[: len(new_indices)]
+            self.topk_index = self.topk_index[: len(new_indices)]
+            self.hidden_states = self.hidden_states[: len(new_indices)]
+            self.verified_id = self.verified_id[: len(new_indices)]
+            if self.allocate_lens is not None:
+                self.allocate_lens = np.asarray(self.allocate_lens)[: len(new_indices)]
+            if self.new_seq_lens is not None:
+                self.new_seq_lens = np.asarray(self.new_seq_lens)[: len(new_indices)]
+        else:
+            self.topk_p = self.topk_p[new_indices]
+            self.topk_index = self.topk_index[new_indices]
+            self.hidden_states = self.hidden_states[new_indices]
+            self.verified_id = self.verified_id[new_indices]
+            if self.allocate_lens is not None:
+                self.allocate_lens = np.asarray(self.allocate_lens)[new_indices]
+            if self.new_seq_lens is not None:
+                self.new_seq_lens = np.asarray(self.new_seq_lens)[new_indices]
+
+    def trim_to_length(self, n: int):
+        self._ensure_host()
+        for f in (
+            "topk_p",
+            "topk_index",
+            "hidden_states",
+            "verified_id",
+            "allocate_lens",
+            "new_seq_lens",
+            "accept_length",
+            "accept_length_cpu",
+            "future_indices",
+        ):
+            v = getattr(self, f, None)
+            if v is not None and len(v) != n:
+                setattr(self, f, np.asarray(v)[:n])
+
+    def merge_batch(self, spec_info: EagleDraftInput):
+        if self.future_indices is not None or spec_info.future_indices is not None:
+            assert self.future_indices is not None and spec_info.future_indices is not None, (
+                "merge_batch requires both EagleDraftInput objects to carry future_indices "
+                "on the relay-buffer path"
+            )
+            self.future_indices = np.concatenate(
+                [np.asarray(self.future_indices), np.asarray(spec_info.future_indices)],
+                axis=0,
+            )
+            if self.allocate_lens is not None and spec_info.allocate_lens is not None:
+                self.allocate_lens = np.concatenate(
+                    [np.asarray(self.allocate_lens), np.asarray(spec_info.allocate_lens)],
+                    axis=0,
+                )
+            else:
+                self.allocate_lens = None
+            if self.new_seq_lens is not None and spec_info.new_seq_lens is not None:
+                self.new_seq_lens = np.concatenate(
+                    [np.asarray(self.new_seq_lens), np.asarray(spec_info.new_seq_lens)],
+                    axis=0,
+                )
+            else:
+                self.new_seq_lens = None
+            if self.accept_length_cpu is not None and spec_info.accept_length_cpu is not None:
+                self.accept_length_cpu = np.concatenate(
+                    [np.asarray(self.accept_length_cpu), np.asarray(spec_info.accept_length_cpu)],
+                    axis=0,
+                )
+            else:
+                self.accept_length_cpu = None
+            return
+
+        if self.hidden_states is None:
+            self.hidden_states = spec_info.hidden_states
+            self.verified_id = spec_info.verified_id
+            self.topk_p = spec_info.topk_p
+            self.topk_index = spec_info.topk_index
+            self.allocate_lens = spec_info.allocate_lens
+            self.new_seq_lens = spec_info.new_seq_lens
+            return
+        if spec_info.hidden_states is None:
+            return
+        self._ensure_host()
+        spec_info._ensure_host()
+        self.hidden_states = np.concatenate([self.hidden_states, spec_info.hidden_states], axis=0)
+        self.verified_id = np.concatenate([self.verified_id, spec_info.verified_id], axis=0)
+        self.topk_p = np.concatenate([self.topk_p, spec_info.topk_p])
+        self.topk_index = np.concatenate([self.topk_index, spec_info.topk_index])
+        self.allocate_lens = np.concatenate([self.allocate_lens, spec_info.allocate_lens])
+        if self.new_seq_lens is not None and spec_info.new_seq_lens is not None:
+            self.new_seq_lens = np.concatenate(
+                [np.asarray(self.new_seq_lens), np.asarray(spec_info.new_seq_lens)]
+            )
+        else:
+            self.new_seq_lens = None
+
+
+@register_pytree_node_class
+@dataclass
+class EagleVerifyInput:
+    """Target-verify input.
+
+    Fully describes token/position/mask/tree-index for verify so
+    ``BaseSpecWorker.verify()`` never reads draft-worker internal state.
+    Under DP (Route 1), per-request fields use DP-padded order; verify
+    metadata must reshape to per-DP view before generating cu_q/kv_lens.
+    """
+
+    # --- Device arrays (enter target verify forward / sampling) ---
+    #: device ``(b*draft_token_num,)`` — flattened draft tokens to verify.
+    draft_token: jax.Array
+    #: device ``(sum(q_i*kv_i),)`` — tree attention mask; shape participates
+    #: in the JIT cache key.  ``None`` when topk=1 (chain mode uses causal).
+    custom_mask: jax.Array | None
+    #: device ``(b*draft_token_num,)`` — verify positions (follows
+    #: ``ForwardBatch`` host/device convention).
+    positions: jax.Array
+    #: device — tree verify index (sampling-kernel convention).
+    retrive_index: jax.Array
+    #: device — tree child pointer for tree sampling.
+    retrive_next_token: jax.Array
+    #: device — tree sibling pointer for tree sampling.
+    retrive_next_sibling: jax.Array
+
+    # --- Static metadata (pytree aux; changes trigger new compile shape) ---
+    spec_steps: int
+    #: per-request verify token count (constant within a precompile shape).
+    draft_token_num: int
+
+    def tree_flatten(self):
+        children = (
+            self.draft_token,
+            self.custom_mask,
+            self.positions,
+            self.retrive_index,
+            self.retrive_next_token,
+            self.retrive_next_sibling,
+        )
+
+        aux_data = {
+            "spec_steps": self.spec_steps,
+            "draft_token_num": self.draft_token_num,
+        }
+        return (children, aux_data)
+
+    @classmethod
+    def tree_unflatten(cls, aux_data, children):
+        obj = cls.__new__(cls)
+        obj.spec_steps = aux_data["spec_steps"]
+        obj.draft_token_num = aux_data["draft_token_num"]
+
+        obj.draft_token = children[0]
+        obj.custom_mask = children[1]
+        obj.positions = children[2]
+        obj.retrive_index = children[3]
+        obj.retrive_next_token = children[4]
+        obj.retrive_next_sibling = children[5]
+
+        return obj
+
+    def prepare_for_verify(self, model_worker_batch: ModelWorkerBatch):
+        sel = model_worker_batch.logits_indices_selector
+        model_worker_batch.seq_lens[sel] = model_worker_batch.seq_lens[sel] - 1
+        model_worker_batch.input_ids = self.draft_token
+        model_worker_batch.positions = self.positions
+        model_worker_batch.return_hidden_states = False
+        model_worker_batch.forward_mode = ForwardMode.TARGET_VERIFY
+        model_worker_batch.spec_info_padded = self
+        model_worker_batch.capture_hidden_mode = CaptureHiddenMode.FULL
+        model_worker_batch.extend_seq_lens = self.draft_token
+
+    def sample(
+        self,
+        model_worker_batch: ModelWorkerBatch,
+        logits_output: LogitsProcessorOutput,
+        rng: nnx.Rngs,
+        mesh: Mesh,
+    ) -> jax.Array:
+        """
+        Verify and find accepted tokens based on logits output and batch
+        (which contains spec decoding information).
+
+        WARNING: This API in-place modifies the states of logits_output
+
+        This API updates values inside logits_output based on the accepted
+        tokens. I.e., logits_output.next_token_logits only contains
+        accepted token logits.
+        """
+        sampling_info = model_worker_batch.sampling_info
+        bs = self.retrive_index.shape[0]
+        if bs != len(sampling_info):
+            sampling_info = copy.deepcopy(sampling_info)
+            # NOTE: retrive_index are the indices of the requests that are kept.
+            # FIXME(pc)
+            sampling_info.filter_batch(np.arange(len(sampling_info)))
+
+        # TODO: support custom sampler, apply the custom logit processors if registered in the sampling info.
+        # if sampling_info.has_custom_logit_processor:
+        #     pass
+        # TODO: Apply penalty
+        # if sampling_info.penalizer_orchestrator.is_required:
+        #     pass
+        # TODO: Apply grammar mask
+        # if vocab_mask is not None:
+        #     pass
+
+        is_all_greedy = sampling_info.is_all_greedy
+        candidates = self.draft_token.reshape(bs, self.draft_token_num)
+
+        if is_all_greedy:
+            with jax.set_mesh(mesh):
+                accept_index, accept_length, predict = verify_tree_greedy(
+                    speculative_num_steps=self.spec_steps,
+                    num_draft_tokens=self.draft_token_num,
+                    draft_tokens=self.draft_token,
+                    retrive_index=self.retrive_index,
+                    retrive_next_token=self.retrive_next_token,
+                    retrive_next_sibling=self.retrive_next_sibling,
+                    next_token_logits=logits_output.next_token_logits,
+                )
+        else:
+            bs = self.retrive_index.shape[0]
+            predict_shape = list(logits_output.next_token_logits.shape)[:-1]
+            predict_shape[-1] += 1
+            predict = jnp.zeros(predict_shape, dtype=jnp.int32)
+
+            accept_index = jnp.full((bs, self.spec_steps + 1), -1, dtype=jnp.int32)
+            accept_length = jnp.zeros((bs,), dtype=jnp.int32)
+            # apply temperature and get target probs
+            expanded_temperature = jnp.repeat(
+                sampling_info.temperatures, self.draft_token_num
+            )  # (bs * draft_token_num, 1)
+            expanded_temperature = jnp.expand_dims(expanded_temperature, axis=-1)
+            target_probs = jax.nn.softmax(
+                logits_output.next_token_logits / expanded_temperature, axis=-1
+            )  # (bs * draft_token_num, vocab_size)
+            target_probs = top_k_renorm_prob(
+                target_probs, jnp.repeat(sampling_info.top_ks, self.draft_token_num)
+            )
+
+            if not jnp.all(sampling_info.top_ps == 1.0):
+                target_probs = top_p_renorm_prob(
+                    target_probs, jnp.repeat(sampling_info.top_ps, self.draft_token_num)
+                )
+
+            # TODO: optimize top_k and top_p by avoiding sort
+            rngs = jax.random.split(rng.params(), 3)
+
+            draft_probs = jnp.zeros(target_probs.shape, dtype=jnp.float32)
+
+            # coins for rejection sampling
+            coins = jax.random.uniform(rngs[1], candidates.shape, dtype=jnp.float32)
+            # coins for final sampling
+            coins_for_final_sampling = jax.random.uniform(rngs[2], (bs,), dtype=jnp.float32)
+            accept_index, accept_length, predict = tree_speculative_sampling_target_only(
+                predicts=predict,
+                accept_index=accept_index,
+                accept_token_num=accept_length,
+                candidates=candidates,
+                retrive_index=self.retrive_index,
+                retrive_next_token=self.retrive_next_token,
+                retrive_next_sibling=self.retrive_next_sibling,
+                uniform_samples=coins,
+                uniform_samples_for_final_sampling=coins_for_final_sampling,
+                target_probs=target_probs,
+                draft_probs=draft_probs,
+                threshold_single=global_server_args_dict["speculative_accept_threshold_single"],
+                threshold_acc=global_server_args_dict["speculative_accept_threshold_acc"],
+                deterministic=True,
+            )
+
+        accept_length = accept_length + 1
+        if SIMULATED_ACCEPTANCE_CONFIG.enabled:
+            target_predict = None
+            if SIMULATED_ACCEPTANCE_CONFIG.token_mode == "real-draft-token":
+                target_predict = jnp.argmax(
+                    logits_output.next_token_logits, axis=-1
+                ).reshape(bs, self.draft_token_num)
+            simulation_rng = jax.random.split(rng.params())[1]
+            accept_index, predict, accept_length = apply_simulated_acceptance(
+                accept_index=accept_index,
+                predict=predict,
+                accept_lens=accept_length,
+                candidates=candidates,
+                target_predict=target_predict,
+                valid_mask=accept_index[:, 0] >= 0,
+                spec_steps=self.spec_steps,
+                topk=model_worker_batch.speculative_eagle_topk,
+                rng=simulation_rng,
+            )
+
+        for arr in (predict, accept_index, accept_length):
+            if hasattr(arr, "copy_to_host_async"):
+                arr.copy_to_host_async()
+        predict = np.asarray(predict)
+        accept_index = np.asarray(accept_index)
+        accept_length = np.asarray(accept_length)
+
+        accept_index = accept_index.flatten()
+        verified_id = np.zeros_like(accept_index, dtype=predict.dtype)
+        verified_id[accept_index != -1] = predict[accept_index[accept_index != -1]]
+        return predict, verified_id, accept_length, accept_index
